JobService: Transition to the Binder Interprocess Communication System
With this transition, we are deprecating the implicit intent
broadcasting system in preparation of Android O.

For more information regarding this API change:
https://commonsware.com/blog/2017/04/11/android-o-implicit-broadcast-ban.html

A brief explanation on Binder (thanks to @Surge1223):

Binder is generally a shorthand for IPC binder, which means interprocess
communication.

This usually involves a client and a service, and communicate through
the kernel via the /dev/binder of the kernel. In order for them to
communicate, the client and remote service needs a way to interface,
which is where the Android Interface Description Language is used.

The client and remote service must both implement an AIDL file,
which generates a stub and proxy that can be used to communicate with
each other.The client usually sends a parcelized stub proxy to
/dev/binder and the service unparcelizes the message via its stub
interface.

Change-Id: I23057a091890614c6d132be52cc6beca43ede200